feat: Modify architecture to be modular
