+++ __init__.py
@@ -6,11 +6,7 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of L2RPN 2023 LJN Agent, a repository for the winning agent of L2RPN 2023 competition. It is a submodule contribution to the L2RPN Baselines repository.
 
-__all__ = ["LJNAgent", "evaluate"]
+__all__ = ["LJNAgent", "evaluate", "LJNAgentTopoNN"]
 
 from l2rpn_baselines.LJNAgent.evaluate import evaluate
-from l2rpn_baselines.LJNAgent.LJNAgent import (
-    L2RPN_IDF_2023_DEFAULT_ACTION_SPACE,
-    L2RPN_IDF_2023_DEFAULT_OPTIM_CONFIG,
-    LJNAgent,
-)
+from l2rpn_baselines.LJNAgent.LJNAgent import LJNAgent, LJNAgentTopoNN
